% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{branching}
\alias{branching}
\title{branching}
\usage{
branching(gen, rv.sim, ...)
}
\arguments{
\item{gen}{generation}

\item{rv.sim}{simulation}

\item{...}{other parameters}
}
\value{
vector
}
\description{
branching
}
