% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchingPlot.R
\name{branchingPlot}
\alias{branchingPlot}
\title{branchingPlot}
\usage{
branchingPlot(gen, rv.sim, ..., reps = 1, logplot = TRUE)
}
\arguments{
\item{gen}{generations}

\item{rv.sim}{simulation of random variables ex. runif(n)}

\item{...}{others parameters}

\item{reps}{number of repetition}

\item{logplot}{if log?}
}
\value{
graphic
}
\description{
branchingPlot
}
