% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSIR.R
\name{computeSIR}
\alias{computeSIR}
\title{computeSIR
Compute susceptible, Infected and Removote at a given time}
\usage{
computeSIR(alpha, beta, initSusc, timeOfSpread)
}
\arguments{
\item{alpha}{infection rate}

\item{beta}{removal rate}

\item{initSusc}{initial susceptible}

\item{timeOfSpread}{time of spread}
}
\value{
vector
}
\description{
computeSIR
Compute susceptible, Infected and Removote at a given time
}
\examples{
computeSIR(0.5, 0.02, 10, 3)
[1]  0 11  0

}
