% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expectedLabouchereSys}
\alias{expectedLabouchereSys}
\title{expectedLabouchereSys
compute the expected winnings per game}
\usage{
expectedLabouchereSys()
}
\value{
numeric
}
\description{
expectedLabouchereSys
compute the expected winnings per game
}
\examples{
> expectedMartingaleSys()
[1] -1.86651
}
