% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expectedMartingaleSys}
\alias{expectedMartingaleSys}
\title{expectedMartingaleSys
compute the expected winnings per game}
\usage{
expectedMartingaleSys()
}
\value{

}
\description{
expectedMartingaleSys
compute the expected winnings per game
}
\examples{
> expectedMartingaleSys()
[1] -2.15629

}
