% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expectedWinBetOnNumber}
\alias{expectedWinBetOnNumber}
\title{expectedWinBetOnNumber
compute the expected winnings per game}
\usage{
expectedWinBetOnNumber()
}
\value{
numeric
}
\description{
expectedWinBetOnNumber
compute the expected winnings per game
}
\examples{
> expectedWinBetOnNumber()
[1] -0.0082

}
