% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expectedWinBetOnRed}
\alias{expectedWinBetOnRed}
\title{expectedWinBetOnRed
compute the expected winnings per game}
\usage{
expectedWinBetOnRed()
}
\value{
numeric
}
\description{
expectedWinBetOnRed
compute the expected winnings per game
}
\examples{
> expectedWinBetOnRed()
[1] -0.02714
}
