% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestFire.R
\name{forestFire}
\alias{forestFire}
\title{forestFire}
\usage{
forestFire(X, apha, beta, pausing = FALSE)
}
\arguments{
\item{X}{data}

\item{apha}{probability to infect on each step}

\item{beta}{probability to be removed}

\item{pausing}{parameter}
}
\description{
forestFire
}
