% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{gameBetOnNumber}
\alias{gameBetOnNumber}
\title{gameBetOnNumber}
\usage{
gameBetOnNumber()
}
\value{
vector
}
\description{
run Betting on a Number
This game involves just one bet. You bet $1 on a particular number, say
17; if the ball lands on that number you win $35, otherwise you lose
}
\examples{
> gameBetOnNumber()
[1] -1  1
}
