% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{gameBetOnRed}
\alias{gameBetOnRed}
\title{gameBetOnRed}
\usage{
gameBetOnRed()
}
\value{
vector
}
\description{
run Betting on Red
This game involves just one bet. You bet $1 on red. If the ball lands on red
you win $1, otherwise you lose.
}
\examples{
> gameBetOnRed()
[1] 1 1
}
