% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{gameLabouchereSys}
\alias{gameLabouchereSys}
\title{gameLabouchereSys}
\usage{
gameLabouchereSys()
}
\value{
vector
}
\description{
Labouchere System
In this game you start with the list of numbers (1, 2, 3, 4). You bet the sum
of the first and last numbers on red (initially $5). If you win you delete the
first and last numbers from the list (so if you win your first bet it becomes
(2,3)), otherwise you add the sum to the end of your list (so if you lose
list is empty, or the bet exceeds $100. If only one number is left on the list,
you bet that number.
}
\examples{
> gameLabouchereSys()
[1] 10  8
}
