% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{gameMartingaleSys}
\alias{gameMartingaleSys}
\title{gameMartingaleSys}
\usage{
gameMartingaleSys()
}
\value{
vector
}
\description{
Martingale System
In this game you start by betting $1 on red. If you lose, you double your
previous bet; if you win, you bet $1 again. You continue to play until you
have won $10, or the bet exceeds $100.
}
\examples{
> gameMartingaleSys()
[1] 10 24
}
