% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameSummary.R
\name{gameSummary}
\alias{gameSummary}
\title{gameSummary
return result of 5 replication}
\usage{
gameSummary()
}
\value{
list
}
\description{
gameSummary
return result of 5 replication
}
\examples{
gameSummary()
[[1]]
[1] -1

[[2]]
[1] 1

[[3]]
[1] -1

[[4]]
[1] 1

[[5]]
[1] 10

[[6]]
[1] 10

[[7]]
[1] 10

[[8]]
[1] 10

[[9]]
[1] 0.4836

[[10]]
[1] 0.4862

[[11]]
[1] 0.02624

[[12]]
[1] 0.02793

[[13]]
[1] 0

[[14]]
[1] 0

[[15]]
[1] 0.95635

[[16]]
[1] 0.95759

[[17]]
[1] 1

[[18]]
[1] 1

[[19]]
[1] 1

[[20]]
[1] 1

[[21]]
[1] 11

[[22]]
[1] 31

[[23]]
[1] 2

[[24]]
[1] 23
}
