% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{neighbours}
\alias{neighbours}
\title{neighbours}
\usage{
neighbours(A, i, j)
}
\arguments{
\item{A}{grid of}

\item{i, j}{coordinate of grid}
}
\value{
number of neighbours
}
\description{
neighbours
}
