% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{propWinBetOnNumber}
\alias{propWinBetOnNumber}
\title{propWinBetOnNumber
compute the proportion of games you win}
\usage{
propWinBetOnNumber()
}
\value{
numeric
}
\description{
propWinBetOnNumber
compute the proportion of games you win
}
\examples{
> propWinBetOnNumber()
[1] 0.02697

}
