% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{propWinBetOnRed}
\alias{propWinBetOnRed}
\title{propWinBetOnRed
compute the proportion of games you win}
\usage{
propWinBetOnRed()
}
\value{
numeric
}
\description{
propWinBetOnRed
compute the proportion of games you win
}
\examples{
propWinBetOnRed()
[1] 0.48502

}
