% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{propWinLabouchereSys}
\alias{propWinLabouchereSys}
\title{propWinLabouchereSys
compute the proportion of games you win}
\usage{
propWinLabouchereSys()
}
\value{
numeric
}
\description{
propWinLabouchereSys
compute the proportion of games you win
}
\examples{
> propWinLabouchereSys()
[1] 0.95786

}
