% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{propWinMartingaleSys}
\alias{propWinMartingaleSys}
\title{propWinMartingaleSys
compute the proportion of games you win}
\usage{
propWinMartingaleSys()
}
\value{
numeric
}
\description{
propWinMartingaleSys
compute the proportion of games you win
}
\examples{
> propWinMartingaleSys()
[1] 0

}
