% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuBranching.R
\name{simuBranching}
\alias{simuBranching}
\title{simuBranching}
\usage{
simuBranching(gen, rv.sim, ...)
}
\arguments{
\item{gen}{number of generation}

\item{rv.sim}{random variable simulation ex. runif(n)}

\item{...}{autheur paramaters}
}
\value{
a numeric
}
\description{
simuBranching
}
