% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuSIR.R
\name{simuSIR}
\alias{simuSIR}
\title{simuSIR}
\usage{
simuSIR(alpha, beta, initSusc, timeOfSpread)
}
\arguments{
\item{alpha}{infection probability}

\item{beta}{removal probability}

\item{initSusc}{initial susceptible}

\item{timeOfSpread}{time of spread}
}
\value{
matrix of SIR in each time (row)
}
\description{
Simulate a Susceptible, Infected and Removed persons in
in such periode and return a matrix in which each row
represente a given time between the beginning of spread and the final
time of spread
}
\examples{
> simuSIR(0.5, 0.03, 100, 10)
[,1] [,2] [,3]
[1,]  100    1    0
[2,]   49   52    0
[3,]    0  100    1
[4,]    0   96    5
[5,]    0   92    9
[6,]    0   87   14
[7,]    0   83   18
[8,]    0   82   19
[9,]    0   82   19
[10,]    0   81   20
[11,]    0   78   23

}
